+++ src/io/build-output.cpp
@@ -115,7 +115,7 @@ vector<OId> monica::parseOutputIds(const J11Array& oidArray)
 {
   vector<OId> outputIds;
 
-  auto getAggregationOp = [](J11Array arr, uint index, OId::OP def = OId::_UNDEFINED_OP_) -> OId::OP
+  auto getAggregationOp = [](J11Array arr, uint8_t index, OId::OP def = OId::_UNDEFINED_OP_) -> OId::OP
   {
     if (arr.size() > index && arr[index].is_string())
     {
@@ -141,7 +141,7 @@ vector<OId> monica::parseOutputIds(const J11Array& oidArray)
   };
 
 
-  auto getOrgan = [](J11Array arr, uint index, OId::ORGAN def = OId::_UNDEFINED_ORGAN_) -> OId::ORGAN
+  auto getOrgan = [](J11Array arr, uint8_t index, OId::ORGAN def = OId::_UNDEFINED_ORGAN_) -> OId::ORGAN
   {
     if (arr.size() > index && arr[index].is_string())
     {
